+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Upcoming
+## 2.5.7
 * Animals now have a small chance to be hostile
   * `enhancedai:can_fight_back` entity type tag has been renamed to `enhancedai:can_be_neutral`
   * New `enhancedai:can_be_hostile` entity type tag